refactor(runtime/llm): extract turn-end stop-gate evaluation to own module
chat-executor-stop-gate-evaluation.ts created (341 LOC) covering the
full once-per-turn stop-hook evaluation sequence:

  - Guard clauses (signal aborted / pending tool protocol / stop
    reason != completed) return { shouldContinueLoop: false } with no
    side effects.
  - Emit the completion_validation_started + per-validator started
    trace envelope that observability clients already consume.
  - When stopHooks are disabled: mark each validator as skipped and
    emit the finished envelope with skipped outcomes.
  - When enabled: call runStopHookPhase, update the runtime-contract
    validator snapshot per hook outcome, and emit the stop_hook
    _execution_finished + completion_validator_finished envelopes.
  - Dispatch prevent_continuation (set stop reason + clear content),
    retry_with_blocking_message (delegate to attemptCompletionRecovery),
    and the fall-through token-budget continuation.

Return value is { shouldContinueLoop: boolean } so the caller owns
the outer do-while continue decision explicitly.

Main loop drops 1,070 -> 789 LOC. Zero behavior delta - the block was
lifted verbatim from the loop with the outer-scope shouldContinueAfterStopGate
flag replaced by the return value. Dead helpers asDelegationOutputValidationCode
and the now-unused BUILTIN_*_HOOK_ID / runStopHookPhase / buildTurnEndStopGateSnapshot
/ updateRuntimeContractValidatorSnapshot imports removed from the main
loop.

Tests green (9 pre-existing failures, zero new regressions).
